Changes in some code and README
Upgraded Readme.md.

Moved the script and style in more relevant folders.

Script no longer imports the CSS as though it were a module and no longer creates the canvas as we now have script link and a pre-existing canvas in the initial HTML page.

Now the orbit object can run in an elliptic orbit.

Made the project compatible with NetBeans as well (by importing it in NetBeans and setting some run configurations to the project).