Adjusted the credits
The credits now are white and have bloom, shown centered in the looking position. The credits will not be generated again and again at each press of the key WITHOUT being deleted as the array was being initialized only once, but not anymore. Also, for the BLOOM_SCENE enum, the CreditsManager has a reference to the object (GameManager) that creates and stores it, via a new parameter in its constructor.